refactor: split RegisterLdtkObjects into LdtkEntityAppExt and LdtkIntCellAppExt
